Correct token handling for macro variable definitions
this way the value of a macro can have multiple tokens. i believe the
simple pointer arithmetic with str?spn would be faster than strtok, but
this doesn't matter.

needed to add a handle in the string fetch routine to handle the newline
only when defining. when we are expanding, we want to echo newlines as
we find them, but the value of a macro shouldn't contain a newline for
it to spit out when it expands. we handle the newline here so htable
doesnt need to worry about them

in that regard, a little TODO: is to fix up the barriers for data
control. what i mean is that htable really shouldnt be worrying about
the contents of the strings it is given to hash, so the interface should
be reworked so that none of that needs to be done in htable.c