Zero print buffer if command fails
as mentioned in the comment above the code, this isn't great becasue we
are discarding the output whenever the command exits iwth anything other
than 0 (which isnt always what the user wants!), but this is the only
way to do this with popen(3). to make it work properly will need a _lot_
more code, but it will be faster (esp. because we wont have to memset
the entire buffer given by getdelim, which defaults to a pretty big one
i think...)